+++ NEWS.md
@@ -1,3 +1,5 @@
+# ypr 0.3.1
+
 - `ypr_populations()` now generates names
 - Fixed Quesnel Lake Trout and Rainbow Trout mix up